+++ modules/module-postgres/src/utils/application-name.ts
@@ -1,0 +1,8 @@
+import { POWERSYNC_VERSION } from '@powersync/service-core';
+
+/**
+ * application_name for PostgreSQL connections to the source database
+ */
+export function getApplicationName() {
+  return `powersync/${POWERSYNC_VERSION}`;
+}