+++ modules/module-mongodb/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @powersync/service-module-mongodb
 
+## 0.10.3
+
+### Patch Changes
+
+- 3e7d629: Fix MongoDB initial replication with mixed \_id types.
+- Updated dependencies [3e7d629]
+- Updated dependencies [e8cb8db]
+  - @powersync/service-core@1.13.3
+
 ## 0.10.2
 
 ### Patch Changes