+++ packages/service-core-tests/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @powersync/service-core-tests
 
+## 0.10.3
+
+### Patch Changes
+
+- Updated dependencies [3e7d629]
+- Updated dependencies [e8cb8db]
+  - @powersync/service-core@1.13.3
+
 ## 0.10.2
 
 ### Patch Changes